`Switched bot uptime calculation to local time`
Changed the method used to calculate the bot's uptime from `datetime.datetime.now(datetime.UTC)` to `datetime.datetime.now()`. This change will cause the bot's uptime to be calculated based on the local time of the system running the bot, rather than UTC. This could lead to differences in the reported uptime depending on the timezone of the system.