Update app version and add webhook naming condition
Updated the application version from "1.8.3" to "1.8.4" in Dockerfile and main.py. This change is reflected in the Dockerfile LABEL and the `bot_version` variable in main.py.

Added a new condition to the `self` function in main.py to prevent the creation of webhooks with names containing the word 'discord' (case insensitive). This is to avoid potential naming conflicts or confusion.