Update Docker build commands and workflow rules in .gitlab-ci.yml
Updated the Docker build commands in the `.gitlab-ci.yml` file. A new builder instance "mybuilder" is created and set as the current builder. Added a `sleep 2` command to ensure previous command completion. Updated `docker buildx build` command with new build arguments, compression, no cache, and multi-platform support. The image is tagged and pushed to the Docker repository. The `after_script` section now removes the "mybuilder" instance. The `timeout` value is set to 90 minutes. The `workflow` rules have been updated to always run if the commit branch is "master".