Update README.md and Docker instructions
Updated the title and content of README.md to better reflect the bot's functionalities and added badges for inviting the bot and displaying the server count. Also, revised the Docker instructions for running the bot and relocated the line about inviting the bot hosted by the author to the introduction section.