Update `name` validation in `create_webhook` function
Updated the condition for checking the validity of the `name` parameter in the `create_webhook` function. The new condition checks if the `name` is an empty string, if its length is less than 1 or greater than 80, or if it is composed only of whitespace characters. If any of these conditions are met, the `name` is set to 'WebhookCreator'. This change ensures compliance with the Discord API's requirements for webhook names.