Update Dockerfile and remove log_handler.py from project
The Dockerfile was updated to use the `python:3.12-alpine`
base image, set the working directory to `/app`, and remove
the `COPY` instruction for the `CustomModules` directory.
Environment variables `TERM=xterm` and `PYTHONUNBUFFERED=1`
were added.

In the `WebhookCreator.pyproj` file, the `<Compile>` entry
for `CustomModules\log_handler.py` was removed, excluding
this file from the project.